Refactoring features code into multiple files
