+++ lhotse/features/__init__.py
@@ -1,0 +1,22 @@
+from .base import (
+    FeatureExtractor,
+    FeatureSet,
+    FeatureSetBuilder,
+    Features,
+    create_default_feature_extractor,
+)
+from .fbank import (
+    Fbank,
+    FbankConfig
+)
+from .mfcc import (
+    Mfcc,
+    MfccConfig
+)
+from .mixer import (
+    FeatureMixer
+)
+from .spectrogram import (
+    Spectrogram,
+    SpectrogramConfig
+)
